backend/requirements.txt
# ============ Web 框架 ============
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ============ HTTP 客户端（调用智谱 AI）============
httpx==0.28.1

# ============ 数据校验 ============
pydantic==2.10.4

# ============ Redis ============
redis[hiredis]==5.2.1

# ============ 环境变量 ============
python-dotenv==1.0.1

# ============ YAML（Prompt 管理）============
pyyaml==6.0.2

# ============ HTML 清洗（微信兼容）============
beautifulsoup4==4.12.3
lxml==5.3.0

# ============ SSE 流式响应 ============
sse-starlette==2.2.1

# ============ 数据库 ============
aiosqlite==0.20.0
